restructure CondTools/SiStrip plugins
