+++ CondTools/SiStrip/plugins/SiStripApvGainBuilder.cc
@@ -1,7 +1,32 @@
-#include "CondTools/SiStrip/plugins/SiStripApvGainBuilder.h"
-#include "CalibTracker/SiStripCommon/interface/SiStripDetInfoFileReader.h"
-#include <iostream>
+// system include files
 #include <fstream>
+#include <iostream>
+
+// user include files
+#include "CalibTracker/SiStripCommon/interface/SiStripDetInfoFileReader.h"
+#include "CommonTools/ConditionDBWriter/interface/ConditionDBWriter.h"
+#include "CondFormats/SiStripObjects/interface/SiStripApvGain.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/FileInPath.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/Exception.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+
+#include "CLHEP/Random/RandFlat.h"
+#include "CLHEP/Random/RandGauss.h"
+
+class SiStripApvGainBuilder : public edm::one::EDAnalyzer<> {
+public:
+  explicit SiStripApvGainBuilder(const edm::ParameterSet& iConfig);
+
+  ~SiStripApvGainBuilder() override = default;
+
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+
+private:
+  const edm::FileInPath fp_;
+  const bool printdebug_;
+};
 
 SiStripApvGainBuilder::SiStripApvGainBuilder(const edm::ParameterSet& iConfig)
     : fp_(iConfig.getUntrackedParameter<edm::FileInPath>("file",
@@ -16,12 +41,16 @@ void SiStripApvGainBuilder::analyze(const edm::Event& evt, const edm::EventSetup
 
   SiStripApvGain obj;
 
+  const auto& reader = SiStripDetInfoFileReader::read(fp_.fullPath());
+  const auto& DetInfos = reader.getAllData();
+
   int count = -1;
-  for (const auto& it : SiStripDetInfoFileReader::read(fp_.fullPath()).getAllData()) {
+  for (const auto& it : DetInfos) {
+    const auto& nAPVs = it.second.nApvs;
     count++;
     //Generate Gain for det detid
     std::vector<float> theSiStripVector;
-    for (unsigned short j = 0; j < it.second.nApvs; j++) {
+    for (unsigned short j = 0; j < nAPVs; j++) {
       float gain = (j + 1) * 1000 + (CLHEP::RandFlat::shoot(1.) * 100);
       if (count < printdebug_)
         edm::LogInfo("SiStripApvGainBuilder") << "detid " << it.first << " \t"
@@ -47,3 +76,8 @@ void SiStripApvGainBuilder::analyze(const edm::Event& evt, const edm::EventSetup
     edm::LogError("SiStripApvGainBuilder") << "Service is unavailable" << std::endl;
   }
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+DEFINE_FWK_MODULE(SiStripApvGainBuilder);
