+++ CondTools/SiStrip/plugins/SiStripCablingTrackerMap.cc
@@ -1,7 +1,31 @@
-#include "CondTools/SiStrip/plugins/SiStripCablingTrackerMap.h"
+#include "FWCore/Utilities/interface/Exception.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "CommonTools/TrackerMap/interface/TrackerMap.h"
+#include "CalibTracker/Records/interface/SiStripDetCablingRcd.h"
+#include "CalibFormats/SiStripObjects/interface/SiStripDetCabling.h"
 #include "CommonTools/TrackerMap/interface/TrackerMap.h"
 
 #include <sstream>
+
+class SiStripCablingTrackerMap : public edm::one::EDAnalyzer<edm::one::WatchRuns> {
+public:
+  SiStripCablingTrackerMap(const edm::ParameterSet& conf);
+  ~SiStripCablingTrackerMap() override = default;
+
+  void beginRun(const edm::Run& run, const edm::EventSetup& es) override;
+  void endJob() override;
+  void endRun(const edm::Run& run, const edm::EventSetup& es) override{};
+  void analyze(const edm::Event& e, const edm::EventSetup& c) override;
+
+private:
+  const edm::ESGetToken<SiStripDetCabling, SiStripDetCablingRcd> detCablingToken_;
+
+  std::unique_ptr<TrackerMap> tkMap_detCab;  //0 for onTrack, 1 for offTrack, 2 for All
+};
 
 SiStripCablingTrackerMap::SiStripCablingTrackerMap(edm::ParameterSet const& conf) : detCablingToken_(esConsumes()) {}
 
@@ -10,7 +34,6 @@ void SiStripCablingTrackerMap::beginRun(const edm::Run& run, const edm::EventSet
 }
 
 //------------------------------------------------------------------------------------------
-
 void SiStripCablingTrackerMap::endJob() {
   tkMap_detCab->save(true, 0, 0, "DetCabling.png");
   tkMap_detCab->print(true, 0, 0, "DetCabling");
@@ -29,3 +52,8 @@ void SiStripCablingTrackerMap::analyze(const edm::Event& e, const edm::EventSetu
     tkMap_detCab->fill(detid, 1);
   }
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+DEFINE_FWK_MODULE(SiStripCablingTrackerMap);
