+++ CondTools/SiStrip/plugins/SiStripPedestalsReader.cc
@@ -1,4 +1,32 @@
-#include "CondTools/SiStrip/plugins/SiStripPedestalsReader.h"
+// system include files
+//#include <memory>
+#include <iostream>
+#include <cstdio>
+#include <sys/time.h>
+
+// user include files
+#include "CondFormats/DataRecord/interface/SiStripPedestalsRcd.h"
+#include "CondFormats/SiStripObjects/interface/SiStripPedestals.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+
+class SiStripPedestalsReader : public edm::one::EDAnalyzer<> {
+public:
+  explicit SiStripPedestalsReader(const edm::ParameterSet&);
+  ~SiStripPedestalsReader() override = default;
+
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+
+private:
+  uint32_t printdebug_;
+  const edm::ESGetToken<SiStripPedestals, SiStripPedestalsRcd> pedestalsToken_;
+};
 
 using namespace std;
 using namespace cms;
@@ -27,3 +55,8 @@ void SiStripPedestalsReader::analyze(const edm::Event& e, const edm::EventSetup&
       }
     }
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+DEFINE_FWK_MODULE(SiStripPedestalsReader);
