+++ CondTools/SiStrip/plugins/SiStripSummaryReader.cc
@@ -1,12 +1,33 @@
+// user include files
+#include "CondFormats/DataRecord/interface/SiStripSummaryRcd.h"
 #include "CondFormats/SiStripObjects/interface/SiStripSummary.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
 
-#include "CondTools/SiStrip/plugins/SiStripSummaryReader.h"
-
+// system include files
 #include <iostream>
 #include <string>
 #include <sstream>
 #include <cstdio>
 #include <sys/time.h>
+
+class SiStripSummaryReader : public edm::one::EDAnalyzer<> {
+public:
+  explicit SiStripSummaryReader(const edm::ParameterSet&);
+  ~SiStripSummaryReader() override = default;
+
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+
+private:
+  uint32_t printdebug_;
+  const edm::ESGetToken<SiStripSummary, SiStripSummaryRcd> summaryToken_;
+};
 
 SiStripSummaryReader::SiStripSummaryReader(const edm::ParameterSet& iConfig)
     : printdebug_(iConfig.getUntrackedParameter<uint32_t>("printDebug", 1)), summaryToken_(esConsumes()) {}
@@ -34,3 +55,8 @@ void SiStripSummaryReader::analyze(const edm::Event& e, const edm::EventSetup& i
     }
   edm::LogInfo("SiStripSummaryReader") << ss.str();
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+DEFINE_FWK_MODULE(SiStripSummaryReader);
