+++ CondTools/SiStrip/plugins/SiStripThresholdReader.cc
@@ -1,5 +1,32 @@
-#include "CondTools/SiStrip/plugins/SiStripThresholdReader.h"
+// user include files
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "CondFormats/SiStripObjects/interface/SiStripThreshold.h"
 #include "CondFormats/DataRecord/interface/SiStripThresholdRcd.h"
+
+// system include files
+#include <memory>
+#include <iostream>
+#include <cstdio>
+#include <sys/time.h>
+
+class SiStripThresholdReader : public edm::one::EDAnalyzer<> {
+public:
+  explicit SiStripThresholdReader(const edm::ParameterSet&);
+  ~SiStripThresholdReader() override = default;
+
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+
+private:
+  uint32_t printdebug_;
+  const edm::ESGetToken<SiStripThreshold, SiStripThresholdRcd> thresholdToken_;
+};
 
 using namespace std;
 using namespace cms;
@@ -39,3 +66,8 @@ void SiStripThresholdReader::analyze(const edm::Event& e, const edm::EventSetup&
       }
     }
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+DEFINE_FWK_MODULE(SiStripThresholdReader);
